Initial (**rough**) implementation of complete account

+++ backend/src/index.ts
@@ -1,11 +1,18 @@
 import boot from "./boot";
 import { Worker } from "worker_threads";
 
-const worker = new Worker(__dirname + "/workers/index");
+let worker: Worker | undefined
+
+console.log(`WORKER: ${process.env.WORKER}`)
+if ((process.env.WORKER || 'true') === 'true') {
+  const worker = new Worker(__dirname + "/workers/index");
+}
 
 const shutdown = async () => {
   console.log("Shutting down gracefully...");
-  await worker.terminate();
+  if (worker) {
+    await worker.terminate();
+  }
   process.exit(0);
 };
